requirements.txt
absl-py==0.12.0
appdirs==1.4.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
crypten==0.1.0
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
entrypoints==0.3
future==0.18.2
google-auth==1.29.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==1.1.1
mistune==0.8.4
msgpack==0.6.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
oauthlib==3.1.0
onnx==1.8.1
packaging==20.3
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
progress==1.5
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.4
testpath==0.4.4
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
